refactor: implement createChatInstance function for managing chat sessions and message handling
